requirements.txt
anyascii==0.3.3
arrow==1.3.0
asgiref==3.9.1
async-timeout==5.0.1
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20241021
django==5.2.6
django-browser-reload==1.21.0
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-cotton==2.1.3
django-extensions==4.1
django-filter==25.1
django-htmx==1.17.0
django-import-export==4.3.8
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-modelcluster==6.4
django-otp==1.6.1
django-permissionedforms==0.1
django-redis==6.0.0
django-seed==0.3.1
django-stubs-ext==5.2.7
django-taggit==6.1.0
django-tailwind-cli==4.4.2
django-tasks==0.8.1
django-treebeard==4.7.1
django-typer==3.3.1
django-unfold==0.62.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
draftjs-exporter==5.1.0
et-xmlfile==2.0.0
faker==37.5.3
filetype==1.2.0
flask==3.1.2
honeypot==0.1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
laces==0.1.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
openpyxl==3.1.5
pillow==11.3.0
pillow-heif==1.1.1
prompt-toolkit==3.0.52
psycopg2-binary==2.9.10
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
qrcode==8.2
redis==6.2.0
requests==2.32.4
rich==14.2.0
semver==3.0.4
shellingham==1.5.4
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
telepath==0.3.1
toposort==1.10
typer-slim==0.19.2
types-python-dateutil==2.9.0.20251008
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.2
wagtail==7.1.1
wcwidth==0.2.14
werkzeug==3.1.3
whitenoise==6.8.2
willow==1.11.0
django-import-export==4.3.8
